Split templater class into mysql and pg templaters
