+++ ossdbtoolsservice/edit_data/templates/__init__.py
@@ -1,0 +1,10 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from ossdbtoolsservice.edit_data.templates.templater import Templater
+from ossdbtoolsservice.edit_data.templates.mysql_templater import MySQLTemplater
+from ossdbtoolsservice.edit_data.templates.pg_templater import PGTemplater
+
+__all__ = ['Templater', 'MySQLTemplater', 'PGTemplater']
